@@ -2,9 +2,7 @@ package gouse
 
 import "testing"
 
-/* Testing min max */
-
-func TestMin(t *testing.T) {
+func TestMinMaxArr(t *testing.T) {
 	arr := []int{1, 2, 3, 4, 5}
 
 	minExpected := 1
@@ -17,8 +15,6 @@ func TestMin(t *testing.T) {
 		t.Errorf("Expected %d, got %d", maxExpected, MaxArr(arr))
 	}
 }
-
-/* Testing intersection */
 
 func TestIntersect(t *testing.T) {
 	t.Parallel()
@@ -71,8 +67,6 @@ func TestIntersect(t *testing.T) {
 		})
 	}
 }
-
-/* Testing callback functions */
 
 func TestIndexBy(t *testing.T) {
 	t.Parallel()
@@ -358,16 +352,416 @@ func TestMapBy(t *testing.T) {
 	}
 }
 
-/* Testing utilities */
-
-func TestExist(t *testing.T) {
-	if !IncludesArr[int]([]int{1, 2, 3}, 1) {
+func TestIncludesArr(t *testing.T) {
+	if !IncludesArr([]int{1, 2, 3}, 1) {
 		t.Errorf("Exist[int]([]int{1, 2, 3}, 1) = false, want true")
 	}
 }
 
 func TestEqual(t *testing.T) {
-	if !IncludesArr[int]([]int{1, 2, 3}, 1) {
+	if !IncludesArr([]int{1, 2, 3}, 1) {
 		t.Errorf("Equal[int](1, 1) = false, want true")
 	}
 }
+
+func TestSumArr(t *testing.T) {
+	arr := []int{1, 2, 3, 4, 5}
+
+	expected := 15
+	if SumArr(arr) != expected {
+		t.Errorf("Expected %d, got %d", expected, SumArr(arr))
+	}
+}
+
+func TestProductArr(t *testing.T) {
+	arr := []int{1, 2, 3, 4, 5}
+
+	expected := 120
+	if ProductArr(arr) != expected {
+		t.Errorf("Expected %d, got %d", expected, ProductArr(arr))
+	}
+}
+
+func TestUniqueArr(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name string
+		arr  []int
+		want []int
+	}{
+		{
+			name: "empty array",
+			arr:  []int{},
+			want: nil,
+		},
+		{
+			name: "no duplicates",
+			arr:  []int{1, 2, 3},
+			want: []int{1, 2, 3},
+		},
+		{
+			name: "duplicates",
+			arr:  []int{1, 2, 2, 3},
+			want: []int{1, 2, 3},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := UniqueArr(tt.arr)
+
+			if len(got) != len(tt.want) {
+				t.Errorf("UniqueArr() = %v, want %v", got, tt.want)
+			}
+
+			for i := range got {
+				if got[i] != tt.want[i] {
+					t.Errorf("UniqueArr() = %v, want %v", got, tt.want)
+				}
+			}
+		})
+	}
+}
+
+func TestMost(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name string
+		arr  []int
+		want int
+	}{
+		// {
+		// 	name: "empty array",
+		// 	arr:  []int{},
+		// 	want: 0,
+		// },
+		{
+			name: "single element",
+			arr:  []int{1},
+			want: 1,
+		},
+		{
+			name: "multiple elements",
+			arr:  []int{1, 2, 2, 3},
+			want: 2,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := Most(tt.arr)
+
+			if got != tt.want {
+				t.Errorf("Most() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestChunk(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name string
+		arr  []int
+		size int
+		want [][]int
+	}{
+		{
+			name: "empty array",
+			arr:  []int{},
+			size: 2,
+			want: nil,
+		},
+		{
+			name: "single element",
+			arr:  []int{1},
+			size: 2,
+			want: [][]int{{1}},
+		},
+		{
+			name: "multiple elements",
+			arr:  []int{1, 2, 3, 4, 5},
+			size: 2,
+			want: [][]int{{1, 2}, {3, 4}, {5}},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := Chunk(tt.arr, tt.size)
+
+			if len(got) != len(tt.want) {
+				t.Errorf("Chunk() = %v, want %v", got, tt.want)
+			}
+
+			for i := range got {
+				if len(got[i]) != len(tt.want[i]) {
+					t.Errorf("Chunk() = %v, want %v", got, tt.want)
+				}
+
+				for j := range got[i] {
+					if got[i][j] != tt.want[i][j] {
+						t.Errorf("Chunk() = %v, want %v", got, tt.want)
+					}
+				}
+			}
+		})
+	}
+}
+
+func TestDiff(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name string
+		a    []int
+		b    []int
+		want []int
+	}{
+		{
+			name: "empty arrays",
+			a:    []int{},
+			b:    []int{},
+			want: nil,
+		},
+		{
+			name: "no difference",
+			a:    []int{1, 2, 3},
+			b:    []int{1, 2, 3},
+			want: nil,
+		},
+		{
+			name: "difference",
+			a:    []int{1, 2, 3},
+			b:    []int{2, 3, 4},
+			want: []int{1},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := Diff(tt.a, tt.b)
+
+			if len(got) != len(tt.want) {
+				t.Errorf("Diff() = %v, want %v", got, tt.want)
+			}
+
+			for i := range got {
+				if got[i] != tt.want[i] {
+					t.Errorf("Diff() = %v, want %v", got, tt.want)
+				}
+			}
+		})
+	}
+}
+
+func TestDrop(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name string
+		arr  []int
+		n    int
+		want []int
+	}{
+		{
+			name: "empty array",
+			arr:  []int{},
+			n:    1,
+			want: nil,
+		},
+		{
+			name: "no drop",
+			arr:  []int{1, 2, 3},
+			n:    0,
+			want: []int{1, 2, 3},
+		},
+		{
+			name: "drop one",
+			arr:  []int{1, 2, 3},
+			n:    1,
+			want: []int{2, 3},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := Drop(tt.arr, tt.n)
+
+			if len(got) != len(tt.want) {
+				t.Errorf("Drop() = %v, want %v", got, tt.want)
+			}
+
+			for i := range got {
+				if got[i] != tt.want[i] {
+					t.Errorf("Drop() = %v, want %v", got, tt.want)
+				}
+			}
+		})
+	}
+}
+
+func TestIndexOfArr(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name  string
+		arr   []int
+		value int
+		want  int
+	}{
+		{
+			name:  "empty array",
+			arr:   []int{},
+			value: 1,
+			want:  -1,
+		},
+		{
+			name:  "not found",
+			arr:   []int{1, 2, 3},
+			value: 4,
+			want:  -1,
+		},
+		{
+			name:  "found",
+			arr:   []int{1, 2, 3},
+			value: 2,
+			want:  1,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := IndexOfArr(tt.arr, tt.value)
+
+			if got != tt.want {
+				t.Errorf("IndexOfArr() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestMerge(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name string
+		arr  [][]int
+		want []int
+	}{
+		{
+			name: "empty array",
+			arr:  [][]int{},
+			want: nil,
+		},
+		{
+			name: "no merge",
+			arr:  [][]int{{1}, {2}, {3}},
+			want: []int{1, 2, 3},
+		},
+		{
+			name: "merge",
+			arr:  [][]int{{1}, {2}, {3}, {4}},
+			want: []int{1, 2, 3, 4},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := Merge(tt.arr...)
+
+			if len(got) != len(tt.want) {
+				t.Errorf("Merge() = %v, want %v", got, tt.want)
+			}
+
+			for i := range got {
+				if got[i] != tt.want[i] {
+					t.Errorf("Merge() = %v, want %v", got, tt.want)
+				}
+			}
+		})
+	}
+}
+
+func TestCompact(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name string
+		arr  []int
+		want []int
+	}{
+		{
+			name: "empty array",
+			arr:  []int{},
+			want: nil,
+		},
+		{
+			name: "no compact",
+			arr:  []int{1, 2, 3},
+			want: []int{1, 2, 3},
+		},
+		{
+			name: "compact",
+			arr:  []int{0, 1, 2, 3},
+			want: []int{1, 2, 3},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := Compact(tt.arr)
+
+			if len(got) != len(tt.want) {
+				t.Errorf("Compact() = %v, want %v", got, tt.want)
+			}
+
+			for i := range got {
+				if got[i] != tt.want[i] {
+					t.Errorf("Compact() = %v, want %v", got, tt.want)
+				}
+			}
+		})
+	}
+}
+
+// func TestSort(t *testing.T) {
+// 	t.Parallel()
+
+// 	tests := []struct {
+// 		name string
+// 		arr  []int
+// 		want []int
+// 	}{
+// 		{
+// 			name: "empty array",
+// 			arr:  []int{},
+// 			want: nil,
+// 		},
+// 		{
+// 			name: "no sort",
+// 			arr:  []int{3, 2, 1},
+// 			want: []int{1, 2, 3},
+// 		},
+// 		{
+// 			name: "sort",
+// 			arr:  []int{3, 1, 2},
+// 			want: []int{1, 2, 3},
+// 		},
+// 	}
+
+// 	for _, tt := range tests {
+// 		t.Run(tt.name, func(t *testing.T) {
+// 			Sort(tt.arr)
+
+// 			for i := range tt.arr {
+// 				if tt.arr[i] != tt.want[i] {
+// 					t.Errorf("Sort() = %v, want %v", tt.arr, tt.want)
+// 				}
+// 			}
+// 		})
+// 	}
+// }
